+++ UriTemplate/VarSpecifier.php
@@ -17,7 +17,7 @@ use League\Uri\Exceptions\SyntaxError;
 use function preg_match;
 
 /**
- * @internal The class exposes the internal representation of an Var Specifier
+ * @internal The class exposes the internal representation of a Var Specifier
  * @link https://www.rfc-editor.org/rfc/rfc6570#section-2.3
  */
 final class VarSpecifier